fix the problem with forwards (haystack index) being broken.
the problem lies in the index -- turns out that forwards is subtly
different from forwards_with_needle_index. it comes down to runs of the
same letter. forwards needs the index to skip to the last letter in a
run, forwards_with_needle_index needs the index to skip to the first
letter in a run.